backend/requirements.txt
fastapi
uvicorn
crewai
langchain_openai
python-dotenv